requirements.txt
streamlit
openai
elevenlabs
python-dotenv
mutagen
ffmpeg-python
requests
Pillow
